[IMM32][SDK][NTUSER] Implement CtfImmGenerateMessage
